+++ experiments/statistics/correlation-cofficient-calculator/src/main.c
@@ -9,8 +9,8 @@
 
 #define DATA_LIST_SIZE 2
 
-int calc_correlation_cofficient_from_data(float *data1, size_t data1_size,
-                                          float *data2, size_t data2_size,
+int calc_correlation_cofficient_from_data(double *data1, size_t data1_size,
+                                          double *data2, size_t data2_size,
                                           size_t data_size);
 
 int main(int argc, char *argv[]) {
@@ -54,22 +54,22 @@ int main(int argc, char *argv[]) {
   return EXIT_SUCCESS;
 }
 
-int calc_correlation_cofficient_from_data(float *data1, size_t data1_size,
-                                          float *data2, size_t data2_size,
+int calc_correlation_cofficient_from_data(double *data1, size_t data1_size,
+                                          double *data2, size_t data2_size,
                                           size_t data_size) {
-  float data1_mean = calc_mean(data1, data1_size);
-  float data2_mean = calc_mean(data2, data2_size);
+  double data1_mean = calc_mean(data1, data1_size);
+  double data2_mean = calc_mean(data2, data2_size);
 
-  log_info("data1_mean: %f", data1_mean);
-  log_info("data2_mean: %f", data2_mean);
+  log_info("data1_mean: %lf", data1_mean);
+  log_info("data2_mean: %lf", data2_mean);
 
-  float *data1_deviations = (float *)malloc(sizeof(float) * data1_size);
+  double *data1_deviations = (double *)malloc(sizeof(double) * data1_size);
   if (data1_deviations == NULL) {
     log_error("Failed to allocate memory for data1_deviations.");
 
     return EXIT_FAILURE;
   }
-  float *data2_deviations = (float *)malloc(sizeof(float) * data2_size);
+  double *data2_deviations = (double *)malloc(sizeof(double) * data2_size);
   if (data2_deviations == NULL) {
     log_error("Failed to allocate memory for data2_deviations.");
 
@@ -79,33 +79,33 @@ int calc_correlation_cofficient_from_data(float *data1, size_t data1_size,
   calc_deviations(data2, data2_size, data2_mean, data2_deviations, data2_size);
 
   for (size_t i = 0; i < data1_size; i++) {
-    log_debug("data1_deviations[%zu]: %f", i, data1_deviations[i]);
+    log_debug("data1_deviations[%zu]: %lf", i, data1_deviations[i]);
   }
   for (size_t i = 0; i < data2_size; i++) {
-    log_debug("data2_deviations[%zu]: %f", i, data2_deviations[i]);
+    log_debug("data2_deviations[%zu]: %lf", i, data2_deviations[i]);
   }
 
-  float data1_variance = calc_variance(data1_deviations, data1_size);
-  float data2_variance = calc_variance(data2_deviations, data2_size);
+  double data1_variance = calc_variance(data1_deviations, data1_size);
+  double data2_variance = calc_variance(data2_deviations, data2_size);
 
-  log_debug("data1_variance: %f", data1_variance);
-  log_debug("data2_variance: %f", data2_variance);
+  log_debug("data1_variance: %lf", data1_variance);
+  log_debug("data2_variance: %lf", data2_variance);
 
-  float data1_std_deviation = calc_std_deviation(data1_variance);
-  float data2_std_deviation = calc_std_deviation(data2_variance);
+  double data1_std_deviation = calc_std_deviation(data1_variance);
+  double data2_std_deviation = calc_std_deviation(data2_variance);
 
-  log_info("data1_std_deviation: %f", data1_std_deviation);
-  log_info("data2_std_deviation: %f", data2_std_deviation);
+  log_info("data1_std_deviation: %lf", data1_std_deviation);
+  log_info("data2_std_deviation: %lf", data2_std_deviation);
 
-  float covariance =
+  double covariance =
       calc_covariance(data1, data1_mean, data2, data2_mean, data_size);
 
-  log_info("covariance: %f", covariance);
+  log_info("covariance: %lf", covariance);
 
-  float correlation_cofficient = calc_correlation_cofficient(
+  double correlation_cofficient = calc_correlation_cofficient(
       covariance, data1_std_deviation, data2_std_deviation);
 
-  log_info("correlation_cofficient: %f", correlation_cofficient);
+  log_info("correlation_cofficient: %lf", correlation_cofficient);
 
   free(data1_deviations);
   free(data2_deviations);
