+++ experiments/statistics/polynomial-approximator/src/calc/calc.c
@@ -5,19 +5,19 @@
 
 #include "log/log.h"
 
-float calc_mean(float *data, size_t data_size) {
-  float sum = 0.0;
+double calc_mean(double *data, size_t data_size) {
+  double sum = 0.0;
   for (size_t i = 0; i < data_size; i++) {
     sum += data[i];
   }
 
-  float mean = sum / data_size;
+  double mean = sum / data_size;
 
   return mean;
 }
 
-int calc_deviations(float *data, size_t data_size, float mean,
-                    float *deviations, size_t deviations_size) {
+int calc_deviations(double *data, size_t data_size, double mean,
+                    double *deviations, size_t deviations_size) {
   for (size_t i = 0; i < data_size && i < deviations_size; i++) {
     deviations[i] = data[i] - mean;
   }
@@ -25,44 +25,45 @@ int calc_deviations(float *data, size_t data_size, float mean,
   return EXIT_SUCCESS;
 }
 
-float calc_variance(float *deviations, size_t deviations_size) {
-  float sum = 0.0;
+double calc_variance(double *deviations, size_t deviations_size) {
+  double sum = 0.0;
   for (size_t i = 0; i < deviations_size; i++) {
     sum += pow(deviations[i], 2);
   }
 
-  float variance = sum / deviations_size;
+  double variance = sum / deviations_size;
 
   return variance;
 }
 
-float calc_std_deviation(float variance) {
-  float std_deviation = sqrt(variance);
+double calc_std_deviation(double variance) {
+  double std_deviation = sqrt(variance);
 
   return std_deviation;
 }
 
-float calc_covariance(float *data1, float data1_mean, float *data2,
-                      float data2_mean, size_t data_size) {
-  float sum = 0.0;
+double calc_covariance(double *data1, double data1_mean, double *data2,
+                       double data2_mean, size_t data_size) {
+  double sum = 0.0;
   for (size_t i = 0; i < data_size; i++) {
     sum += (data1[i] - data1_mean) * (data2[i] - data2_mean);
   }
 
-  float covariance = sum / data_size;
+  double covariance = sum / data_size;
 
   return covariance;
 }
 
-float calc_correlation_cofficient(float covariance, float data1_std_deviation,
-                                  float data2_std_deviation) {
-  float correlation_cofficient =
+double calc_correlation_cofficient(double covariance,
+                                   double data1_std_deviation,
+                                   double data2_std_deviation) {
+  double correlation_cofficient =
       covariance / (data1_std_deviation * data2_std_deviation);
 
   return correlation_cofficient;
 }
 
-int solve_linear_system(float **matrix, float *vector, float *solution,
+int solve_linear_system(double **matrix, double *vector, double *solution,
                         size_t size) {
   for (size_t k = 0; k < size - 1; k++) {
     if (matrix[k][k] == 0.0) {
@@ -72,7 +73,7 @@ int solve_linear_system(float **matrix, float *vector, float *solution,
     }
 
     for (size_t i = k + 1; i < size; i++) {
-      float factor = matrix[i][k] / matrix[k][k];
+      double factor = matrix[i][k] / matrix[k][k];
 
       for (size_t j = k; j < size; j++) {
         matrix[i][j] -= factor * matrix[k][j];
@@ -89,7 +90,7 @@ int solve_linear_system(float **matrix, float *vector, float *solution,
       return EXIT_FAILURE;
     }
 
-    float sum = 0.0;
+    double sum = 0.0;
     for (size_t j = i + 1; j < size; j++) {
       sum += matrix[i][j] * solution[j];
     }
@@ -100,25 +101,25 @@ int solve_linear_system(float **matrix, float *vector, float *solution,
   return EXIT_SUCCESS;
 }
 
-float calc_r_squared(float *observed, float *predicted, int n) {
-  float sum_observed = 0.0;
+double calc_r_squared(double *observed, double *predicted, int n) {
+  double sum_observed = 0.0;
   for (int i = 0; i < n; i++) {
     sum_observed += observed[i];
   }
 
-  float mean_observed = sum_observed / n;
+  double mean_observed = sum_observed / n;
 
-  float sum_squared_errors = 0.0;
-  float sum_squared_total = 0.0;
+  double sum_squared_errors = 0.0;
+  double sum_squared_total = 0.0;
   for (int i = 0; i < n; i++) {
-    float error = observed[i] - predicted[i];
+    double error = observed[i] - predicted[i];
     sum_squared_errors += error * error;
 
-    float total_deviation = observed[i] - mean_observed;
+    double total_deviation = observed[i] - mean_observed;
     sum_squared_total += total_deviation * total_deviation;
   }
 
-  float r_squared = 1 - (sum_squared_errors / sum_squared_total);
+  double r_squared = 1 - (sum_squared_errors / sum_squared_total);
 
   return r_squared;
 }
