+++ experiments/statistics/polynomial-approximator/src/calc/calc.h
@@ -3,18 +3,18 @@
 
 #include <stddef.h>
 
-extern float calc_mean(float *data, size_t data_size);
-extern int calc_deviations(float *data, size_t data_size, float mean,
-                           float *deviations, size_t deviations_size);
-extern float calc_variance(float *deviation, size_t deviation_size);
-extern float calc_std_deviation(float variance);
-extern float calc_covariance(float *data1, float data1_mean, float *data2,
-                             float data2_mean, size_t data_size);
-extern float calc_correlation_cofficient(float covariance,
-                                         float data1_std_deviation,
-                                         float data2_std_deviation);
-extern int solve_linear_system(float **matrix, float *vector, float *solution,
+extern double calc_mean(double *data, size_t data_size);
+extern int calc_deviations(double *data, size_t data_size, double mean,
+                           double *deviations, size_t deviations_size);
+extern double calc_variance(double *deviation, size_t deviation_size);
+extern double calc_std_deviation(double variance);
+extern double calc_covariance(double *data1, double data1_mean, double *data2,
+                             double data2_mean, size_t data_size);
+extern double calc_correlation_cofficient(double covariance,
+                                         double data1_std_deviation,
+                                         double data2_std_deviation);
+extern int solve_linear_system(double **matrix, double *vector, double *solution,
                                size_t size);
-extern float calc_r_squared(float *observed, float *predicted, int n);
+extern double calc_r_squared(double *observed, double *predicted, int n);
 
 #endif
